Clean up warnings emissions during tests.
- Stop using the PID-version of subscribe/unsubscribe where possible
- Expose undocumented `call` for the remaining case
- Add explicit test for the deprecation warnings

I tried to fix the `h` documentation for `subscribe` and `unsubscribe`
to only show the non-deprecated versions, but failed in the `subscribe`
case. I think `unsubscribe` might have worked because of the differing
arities between recommended and deprecated. Rather than have it
half-done I'll leave it to someone more familiar.