Allow getting tracker data by key
`Phoenix.Tracker.get_by_key` allows us to request a specific piece of
data instead of traversing the whole list. I thought I would return an
empty map if we do not find the key in the list because it seemed like a
sane default to me. When I ask for a map I would like to receive one.

Amos King @adkron <amos@binarynoggin.com>