Fix zombie nodes if a node has not been seen yet
Previously with three nodes, n1, n2 and n3 in a cluster, when n2 went
down before it had been seen by n3, but after it has been seen by n1,
the pid would be sent to n3 when the CRDT was synced (from n1).

This would result in a zombie pid being present in n3, but not in n1.
When n2 came back up, it was possible that the zombie pid from its past
life would be in its own list.

To fix this, the ups are now detected after a merge, and the heartbeat
handler manually triggered for each up node. This ensures that the
trackers observe the down locally, even if they have never received the
heartbeat.